_requirements.txt
cv2==4.7.0.72
gym==0.21.0
matplotlib==3.7.1
numpy==1.24.2
optuna==3.1.1
pandas==2.0.0
Pillow==9.5.0
python-dotenv==1.0.0
retro==unknown
stable-baselines3==1.8.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
torch==2.0.0
tqdm==4.65.0